% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xhelp.functions.R
\name{R2nls}
\alias{R2nls}
\title{R2NLS function for calculating R2 for non-linear relationship}
\usage{
R2nls(nls.obj, fitdata)
}
\description{
R2NLS function for calculating R2 for non-linear relationship
}
