% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.assemble.negstandard.R
\name{X.assemble.negstandard}
\alias{X.assemble.negstandard}
\title{X.assemble.negstandard}
\usage{
X.assemble.negstandard(data = NULL, sep = ";", cutoff = 0)
}
\arguments{
\item{data}{Data frame with two columns: First for protein IDs and second for separated GO terms.}

\item{sep}{Separator for the GO terms in the second column}

\item{cutoff}{Numeric: How many GO terms can a pair of proteins share to still be included in the list of 
non-interacting proteins?}
}
\value{
A data frame with pair-wise negative standard table for protein-protein interactions.
}
\description{
Using a database table with Entry names and GO terms, assemble a table of putatively non-interacting proteins.
}
\examples{
GSneg <- X.assemble.negstandard(data=GOterms, sep=";")
}
