% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.assemble.posstandard.R
\name{X.assemble.posstandard}
\alias{X.assemble.posstandard}
\title{X.assemble.posstandard}
\usage{
X.assemble.posstandard(standard = NULL, sep = ";")
}
\arguments{
\item{standard}{Data frame with two columns: First for the complex name and second for comma-separated protein subunit entries.}

\item{sep}{Separator that separates protein IDs in the second column.}
}
\value{
A data frame with pair-wise positive standard table for protein-protein interactions.
}
\description{
Using a database table with Entry names and GO terms, assemble a table of putatively non-interacting proteins.
}
\examples{
GSpos <- X.assemble.posstandard(gold.standard, sep=";")
}
