% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.assemble.traindata.R
\name{X.assemble.traindata}
\alias{X.assemble.traindata}
\title{X.assemble.traindata}
\usage{
X.assemble.traindata(data = NULL, standard = NULL)
}
\arguments{
\item{data}{Data frame with columns 'protein1', 'protein2' and other predictor columns.}

\item{standard}{Data frame with columns 'protein1', 'protein2' and 'complex'. The column complex should have values of 1 or 0
for interacting and non-interacting proteins, respectively.}
}
\value{
A data frame with complex labels from standard assigned to observed data.
}
\description{
Using available data and the gold standard, assemble training set for the model training.
}
\examples{
GS_specific <- X.assemble.traindata(all.predictors[[tp]][[rep]],GS)
}
