% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.average.crossmodels.R
\name{X.average.crossmodels}
\alias{X.average.crossmodels}
\title{X.average.crossmodels}
\usage{
X.average.crossmodels(
  data = NULL,
  scores.col = NA,
  labels.col = NA,
  evaluate = FALSE,
  standard.set = NULL,
  eval.metric = "prc"
)
}
\arguments{
\item{data}{Data frame with columns for protein1, protein2 and columns 'score', 'cycle' and 'split'.}

\item{scores.col}{Character string: Name of the columns with prediction values.}

\item{labels.col}{Character string: Name of the columns with labels.}

\item{evaluate}{Logical: Should the averaged predictions be evaluated? Default is FALSE. If TRUE, standard.set must be included}

\item{standard.set}{Dataframe with columns protein1, protein2 and another column with labels.}

\item{eval.metric}{Character string: Evaluation metric to be used for evaluating the model performance. Options same as in X.evaluate.}
}
\value{
A list with four elements: $data is the data.frame with averaged predictions, $eval.metric is the integer specifying the
resulting evaluation metric, $metric.plotdata are underlying data for plotting the evaluation metric and $plot is the plot itself.
}
\description{
Average cross-trained models to get final prediction values for one dataset.
}
\examples{
comp.models[[rep]] <- X.average.crossmodels(data=cross.model$data, eval.metric="prc", evaluate=TRUE, standard.set=GS_specific)
}
