% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.average.reps.R
\name{X.average.reps}
\alias{X.average.reps}
\title{X.average.reps}
\usage{
X.average.reps(
  data = NULL,
  scores.col = NA,
  weights = NULL,
  evaluate = FALSE,
  standard.set = NULL,
  labels.col = NA,
  eval.metric = "prc"
)
}
\arguments{
\item{data}{List of data frame with columns for protein1, protein2 and a column with predictions.}

\item{scores.col}{Character string: Name of the column with prediction values.}

\item{weights}{Numeric vector: Typically values between 0-1, giving weights for each replicate for weighted averaging.}

\item{evaluate}{Logical: Should the averaged predictions be evaluated? Default is FALSE. If TRUE, standard.set must be included}

\item{standard.set}{Dataframe with columns protein1, protein2 and another column with labels.}

\item{labels.col}{Character string: Name of the standard.set column with labels.}

\item{eval.metric}{Character string: Evaluation metric to be used for evaluating the model performance. Options same as in X.evaluate.}
}
\value{
A list with four elements: $data is the data.frame with averaged predictions, $eval.metric is the integer specifying the
resulting evaluation metric, $eval.data are underlying data for plotting the evaluation metric and $plot is the plot itself.
}
\description{
Average predictions for multiple models across replicates.
}
\examples{
final.model <- X.average.reps(data=comp.models.data, evaluate=TRUE,standard.set=GS)
}
