% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.calculate.cors.R
\name{X.calculate.cors}
\alias{X.calculate.cors}
\title{X.calculate.cors}
\usage{
X.calculate.cors(data = NULL, cor.type = "pearson")
}
\arguments{
\item{data}{Data frame with one column per temperature and protein IDs in row names.}

\item{cor.type}{Type argument in Hmisc::rcorr}
}
\value{
Data frame with pairwise protein interactions and R2 values
}
\description{
Using one of the network building algorithms, cluster initial network into individual protein complexes.
}
\examples{
all.predictors <- X.calculate.cors(MCdata.scaled$data)
}
