% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.calculate.predictors.R
\name{X.calculate.predictors}
\alias{X.calculate.predictors}
\title{X.calculate.predictors}
\usage{
X.calculate.predictors(data = NULL, features = NULL, funs = c(), prefixes = NA)
}
\arguments{
\item{data}{Data frame with a column 'protein' and other columns with features.}

\item{features}{Vector of character strings: Names of columns with features. Default is NULL, in which case all but the protein column will be taken.}

\item{funs}{Vector of character strings: Functions to be used for calculation of predictors, matching the order of predictors given in vars.}

\item{prefixes}{Vector of character strings: Prefixes for naming of new columns in the new table. One for each var.}
}
\value{
Data frame with all predictors for pairwise protein interactions.
}
\description{
Using one of the network building algorithms, cluster initial network into individual protein complexes.
}
\examples{
feature.predictors <- X.calculate.predictors(data=features, 
      features=c("Ti","logAUC","logABL","logLoss","Penalty_trans"), 
      funs=c(rep("absdif",4),"Xsum"),
      prefixes=c(rep("d",4),"sum")
    )
}
