% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.calibrate.scores.R
\name{X.calibrate.scores}
\alias{X.calibrate.scores}
\title{X.calibrate.scores}
\usage{
X.calibrate.scores(
  data = NULL,
  standard.set = NULL,
  scores.col = NA,
  labels.col = NA
)
}
\arguments{
\item{data}{Data frame with columns for protein1, protein2 and a column with scores.}

\item{standard.set}{Data frame with column for protein1, protein2 and a column with labels.}

\item{scores.col}{Name of the data column with model scores.}

\item{labels.col}{Name of the standard.set column with labels with values 0 for non-interacting and 1 for interacting proteins.}
}
\value{
A list of three: $data with the ata frame with the original data and an additional column 'probabilities' with the calibrated probabilities,
$plot with the calibration plot and $model with the model fitted into the original data.
}
\description{
Calibrate model scores with sigmoidal model to approximate interaction probabilities
}
\examples{
calibrated.model <- X.calibrate.scores(final.model$data, GS)

}
