% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.cluster.complexes.R
\name{X.cluster.complexes}
\alias{X.cluster.complexes}
\title{X.cluster.complexes}
\usage{
X.cluster.complexes(data = NULL, scores.col = NA)
}
\arguments{
\item{data}{Data frame with pair-wise interactions (columns 'protein1' and 'protein2') and a column with probabilities from ML model.}

\item{scores.col}{Character string: Name of the columns with prediction values.}
}
\description{
Internal function with simple clustering algorithm.
}
\examples{
clustered <- X.cluster.complexes(data=cut_predictions, scores.col="score")
}
