% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.complexes.to.pairwise.R
\name{X.complexes.to.pairwise}
\alias{X.complexes.to.pairwise}
\title{X.complexes.to.pairwise}
\usage{
X.complexes.to.pairwise(complexes = list())
}
\arguments{
\item{complexes}{A list of vectors where each vector represents subunits of a protein complex.}
}
\value{
A data.frame with pair-wise interaction table.
}
\description{
Transform a list of protein complexes into a pairwise interaction table.
}
\examples{
pairwise.table <- X.complexes.to.pairwise(list_of_complexes)
}
