% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.crosstrain.tree.R
\name{X.crosstrain.tree}
\alias{X.crosstrain.tree}
\title{X.crosstrain.tree}
\usage{
X.crosstrain.tree(
  data = NULL,
  standard.set = NULL,
  train.cycles = 25,
  train.split = 4,
  labels.col = NA,
  tree.type = "RF",
  evaluate = FALSE,
  eval.metric = "prc",
  plot = TRUE,
  train.pars = list(),
  save.dir = "models",
  save.output = c("models", "plots")
)
}
\arguments{
\item{data}{Data frame with all data coming from an experiment. Each row corresponds to a unique pair of proteins and
other columns must include all predictors that are in the standard set.}

\item{standard.set}{Data frame with predictors for pairs of proteins and with labels.}

\item{train.cycles}{Integer: How many times should the training be repeated? Default is 25.}

\item{train.split}{Integer: To how many individual training sets should the data be split in each training cycle? Default is 4.}

\item{labels.col}{Character string: 'data' column name with labels (1 for complex-forming, 0 for others.) Default is NA, in which case
the last column will be considered as the column with labels.}

\item{tree.type}{Character string: What type of classification tree should be used for the model training? Options are "J48","CART", "PART", "C5.0","RF".}

\item{evaluate}{Logical: Should each model be evaluated? Default is FALSE. If TRUE, evaluation plots will be saved in the save.dir and}

\item{eval.metric}{Character string: What type of evaluation metric should be used for evaluation? Options as in X.evaluate
a table with evaluation metrics will be outputted.}

\item{plot}{Logical: If evaluate is TRUE, should the plots be saved?}

\item{train.pars}{Named list: Tuning parameters for classification tree training. Default is list(), in which case the default tuning parameters
as specified in the arguments of the function X.model.tree will be used.}

\item{save.dir}{Character string: Name of the folder to be used to save the models. Default is "models".}
}
\value{
A list with three elements. $data for all prediction scores, $metric.data for all underlying data for evaluation
metric calculation, $metric.plots for metric plots.
}
\description{
Train classification tree model with a cross-training algorithm.
}
\examples{
cross.model <- X.crosstrain.tree(data=all.predictors[[tp]][[rep]],standard.set=GS_specific,train.pars=best.train.pars,
        evaluate=TRUE,plot=TRUE,
        train.split=3, train.cycles=10)
}
