% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.extract.raw.features.R
\name{X.extract.raw.features}
\alias{X.extract.raw.features}
\title{X.extract.raw.features}
\usage{
X.extract.raw.features(data = NULL, abundances = c(1, 2), remains = 10)
}
\arguments{
\item{data}{Data frame with data containing a column 'protein' for protein IDs, 'length' for protein sequence length and
one column for abundances at each temperature labeled Ab1-AbN.}

\item{abundances}{Vector of integers: Which temperatures should be used to estimate protein abundances?.}

\item{remains}{Vector of integers: Which temperatures should be used to estimate protein remains?}
}
\value{
A data frame with extracted features.
}
\description{
Extract features for model training from raw data.
}
\examples{
features.raw <- X.extract.raw.features(MCdata.clean$data, abundances=c(1,2), remains=10)
}
