% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.fit.meltcurves.R
\name{X.fit.meltcurves}
\alias{X.fit.meltcurves}
\title{X.fit.meltcurves}
\usage{
X.fit.meltcurves(data = NULL)
}
\arguments{
\item{data}{Data frame with one column for each temperature, starting with 'T' and followed by number and a column 'protein' with protein IDs.}
}
\value{
A list with four elements: $data with fitted data parameters, $all.data with connected table of input and output data,
$fits with nls fit objects for each fitted protein and $plot showing the distribution of these parameters across dataset.
}
\description{
Fit scaled protein melting curves to output fitting parameters.
}
\examples{
MC.data.fitted <- X.fit.meltcurves(MC.data.scaled$data)
}
