% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.interconnect.network.R
\name{X.interconnect.network}
\alias{X.interconnect.network}
\title{X.interconnect.network}
\usage{
X.interconnect.network(data = NULL, scores.col = NA)
}
\arguments{
\item{data}{A data.frame with pair-wise interaction table.}

\item{scores.col}{Character string: Name of the columns with prediction values.}
}
\value{
A data.frame with pair-wise interaction table.
}
\description{
With pairwise interaction table as an input, interconnect protein complexes such that each two subunits in the same cluster whose pairwise
interaction does not pass the cutoff is considered as an interaction.
}
\examples{
pairwise.table <- X.interconnect.network(averaged_predictions$predicted.dataset \%>\% filter(pred>=sug_cutoff))
}
