% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.network.stats.R
\name{X.network.stats}
\alias{X.network.stats}
\title{X.network.stats}
\usage{
X.network.stats(
  data = NULL,
  scores.col = NA,
  standard.set = NULL,
  labels.col = NA,
  complex.stats = FALSE,
  interconnected = FALSE
)
}
\arguments{
\item{data}{Data frame with columns 'protein1' and 'protein2' and pairwise prediction values.}

\item{scores.col}{Character string: Name of the data column with model scores.}

\item{standard.set}{Data frame with columns protein1, protein2 and another column with labels.}

\item{labels.col}{Character string: Name of the columns with labels.}

\item{complex.stats}{Logical: Should the statistics specific for protein complexes be calculated? Do not use with unclustered networks. Default is FALSE.}

\item{interconnected}{Logical: If TRUE, the statistics will be calculated on interconnected network, meaning that even the subunits
that are present in the same cluster, but did not pass the prediction cut-off, will be considered interconnected. Default is FALSE.}
}
\value{
A list with two elements: $data with underlying data and $plot with the plotted statistics.
}
\description{
Evaluate a network of protein complexes
}
\examples{
network.stats <- X.network.stats(data)
}
