% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.pairwise.to.complexes.R
\name{X.pairwise.to.complexes}
\alias{X.pairwise.to.complexes}
\title{X.pairwise.to.complexes}
\usage{
X.pairwise.to.complexes(data = NULL, names.col = NA, scores.col = NULL)
}
\arguments{
\item{data}{A data.frame with pairwise interaction table with column names 'protein1' and 'protein2'.}

\item{names.col}{Character string: Name of the column with complex names.}
}
\value{
A named list with complex names as the list names and subunits as vector elements.
}
\description{
Transform a pairwise interaction table into a list of protein complexes.
}
\examples{
pairwise.table <- X.complexes.to.pairwise(list_of_complexes)
}
