% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.PCA.reduction.R
\name{X.PCA.reduction}
\alias{X.PCA.reduction}
\title{X.PCA.reduction}
\usage{
X.PCA.reduction(
  data = NULL,
  average = FALSE,
  plot = FALSE,
  separate.plots = TRUE,
  plot.complexes = NA,
  plot.colors = NA,
  plot.color.by = "condition",
  plot.path = FALSE,
  plot.animate = NA
)
}
\arguments{
\item{data}{Data frame with features. Columns are 'protein', 'condition', 'replicate' and feature columns.}

\item{average}{Logical: Should the replicates be averaged before the PCA? Default is FALSE.}

\item{plot}{Logical: Should the results be plotted?}

\item{separate.plots}{Logical: Should the plots be plotted separately for different conditions? Default is FALSE.}

\item{plot.complexes}{Named list of vectors of complexes.}

\item{plot.colors}{Vector: Colors to be used for plotting}

\item{plot.path}{Logical: Should the same proteins by connected between the conditions?}

\item{plot.animate}{Character string: 'Time' for time-based animation and 'State' for state-based animation. Default is NA for no animation.}

\item{plot.colors.by}{Character string: By which column should the plots be colored? Default is 'condition'.}
}
\value{
List of five elements: $data contains PCA coordinates for each protein, $PCA contains information about the PCA,
var.explained shows what percentage of variation is explained by which PC, $plots are PCA plots and $separate.plots is
a list of separate PCA plots.
}
\description{
Examine state of an assumed complex on a PCA plot using extracted features
}
\examples{
data.PCAs <- X.PCA.reduction(all.features.data,plot=TRUE,plot.complexes=complexes)
}
