% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.plot.meltcurves.R
\name{X.plot.meltcurves}
\alias{X.plot.meltcurves}
\title{X.plot.meltcurves}
\usage{
X.plot.meltcurves(
  data = NULL,
  fits = NULL,
  samples = NA,
  replicates = NA,
  color.col = "Sample",
  pointshape.col = "Replicate",
  colors = "rainbow",
  points = c(16, 17, 18, 8, 4, 3, 9, 10, 11, 12, 13, 14),
  fit.length = 100,
  pdf.export = TRUE,
  pdf.folder = ".",
  pdf.name = "MCcurves"
)
}
\arguments{
\item{data}{Data frame or list of data frames: Scaled data with removed outliers and fitting statistics, ideally $data output from X.scale.meltcurves.}

\item{fits}{List containing all fit objects, ideally $fits output from X.meltcurve.fit. The list must be named and the names must correspond to id column
values of the data. Alternatively, a list of such lists of same length as the lenght of the list of data.}

\item{samples}{Character vector: names of samples in the same order as the list of data. NA if only one data frame.}

\item{replicates}{Character vector: names of replicates in the same order as the list of data. NA if only one data frame.}

\item{color.col}{Character string: Which column of the data frame(s) should be used for color scaling?}

\item{pointshape.col}{Character string: Which column of the data frame(s) should be used for point shape definition?}

\item{colors}{Character vector: "rainbow" will plot curves from red to violet across sample. "distinct" will use preset distinct color scheme.
User can also define a vector of colors.}

\item{points}{Integer vector: numbers that define point types used for plotting.}

\item{fit.length}{Integer: How many points should be used for fitting curves. Default is 100 which is sufficient for plotting.}

\item{pdf.export}{Logical: If TRUE (default), a pdf with all plots will be exported.}

\item{pdf.folder}{Character string: Name of the directory for pdf export. Default is the working directory.}

\item{pdf.name}{Character string: Name of the exported pdf file. Default is 'MCcurves'}
}
\value{
A list of melting curve plots.
}
\description{
Plots fitted melting curves.
}
\examples{
MCdata.plotted <- X.plot.meltcurves(all.sdata \%>\% unlist(recursive=FALSE),all.fits \%>\% unlist(recursive=FALSE), samples=rep(c("tp28","tp40"),each=3),replicates=c(1,2,3,1,2,3))
}
