% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.postprocess.R
\name{X.postprocess}
\alias{X.postprocess}
\title{X.postprocess}
\usage{
X.postprocess(
  data = NULL,
  mode = "trim",
  scores.col = NA,
  weighted = FALSE,
  zcut = 2,
  init.stats = FALSE,
  final.stats = FALSE,
  standard.set = NULL,
  labels.col = NA
)
}
\arguments{
\item{data}{Data frame with pair-wise interactions (columns 'protein1' and 'protein2') and a column with probabilities from ML model.}

\item{mode}{Character string:  'trim' for trimming of not-well-connected proteins and 'split' for separating of two loosely connected subunits.
Default is 'trim'.}

\item{scores.col}{Character string: Name of the columns with prediction values.}

\item{zcut}{Numeric: Sub-units with z-score above this value will be removed. Default is 2.}

\item{init.stats}{Logical: Should the initial network stats be calculated? Default is FALSE.}

\item{final.stats}{Logical: Should the final network stats be calculated? Default is FALSE.}

\item{standard.set}{Data frame with columns protein1, protein2 and another column with labels.}

\item{labels.col}{Character string: Name of the columns with labels.}

\item{weight}{Logical: Should the prediction values be taken into account for the postprocessing? Default is FALSE}
}
\value{
A list with three elements: $data with columns 'protein1', 'protein2' and score contains all data for plotting or further processing
of the network, 'stats_initial' contains statistics of the network before clustering. and 'stats_final' contains statistics of the network after
clustering.
}
\description{
Using one of the network building algorithms, cluster initial network into individual protein complexes.
}
\examples{
network.post.split <- X.postprocess(data=network.refined$data, mode="split", scores.col="score", final.stats=TRUE, 
        standard.set=GS,labels="complex", weighted=FALSE)

network.post.trim <- X.postprocess(data=network.post.split$data, mode="trim", scores.col="score", final.stats=TRUE, 
        standard.set=GS,labels="complex", weighted=TRUE)
}
