% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.predict.moonlighters.R
\name{X.predict.moonlighters}
\alias{X.predict.moonlighters}
\title{X.predict.moonlighters}
\usage{
X.predict.moonlighters(
  data = NULL,
  complexes = list(),
  plot = TRUE,
  sig.cutoff = 0.05,
  moon.proportion.cutoff = 1/3,
  min.reps = 1,
  min.conditions = 1,
  noPCs = 0,
  weights = NULL,
  kmeans.maxiter = 10,
  kmeans.nstart = 5,
  p.adj.method = "none"
)
}
\arguments{
\item{data}{Data frame with features reduce by PCA. Columns are 'protein', and PC columns starting with 'PC'. Additional columns
'condition' and 'replicate' can be used.}

\item{complexes}{Named list: A list where each element is vector with proteins and the name of that element is the name of the protein complex.}

\item{plot}{Logical: Should the plots be plotted?}

\item{sig.cutoff}{Numeric: Significance level cutoff, 0.05 by default.}

\item{moon.proportion.cutoff}{Upper proportion of moonlighting subunits cutoff. Default is 1/3.}

\item{min.reps}{Integer: At least in how many replicates should a subunit be significantly moonlighting to appear in hits?}

\item{min.conditions}{Integer: At least in how many conditions should a subunit be significantly moonlighting to appear in hits?}

\item{noPCs}{Integer: How many principal components should be used for the clustering?}

\item{weights}{Numeric vector: weights for each principal component for k.means clustering.
the complex centroid, distance from which is further calculated and gives}

\item{kmeans.maxiter}{Integer: Maximum interation for kmeans clustering.}

\item{kmeans.nstart}{Integer: How many starting sets should be tried for kmeans clustering?}

\item{p.adj.method}{Character string: One of the p.adj methods, see ?p.adjust. Default is "none".}
}
\value{
PCA data and different supporting information.
}
\description{
Calculate assembly factors from complex features reduce into a PCA plot.
}
\examples{
data.moonlighters <- X.predict.moonlighters(data=data.PCAs$data, min.reps=2)
}
