% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.refine.complexes.R
\name{X.refine.complexes}
\alias{X.refine.complexes}
\title{X.refine.complexes}
\usage{
X.refine.complexes(
  data = NULL,
  algo = "SCBP",
  scores.col = NA,
  init.stats = FALSE,
  final.stats = FALSE,
  standard.set = NULL,
  labels.col = NA,
  rep.steps = 100,
  SP.shift = 0.2,
  SR.rows = 20
)
}
\arguments{
\item{data}{Data frame with pair-wise interactions (columns 'protein1' and 'protein2') and a column with probabilities from ML model.}

\item{algo}{Character string: What algorithm should be used for refining? Options are 'SCBR' for shuffling-complexes-by-rows or
'SCBP' for shuffling complexes by prediction value. Default is 'SCBP'.}

\item{scores.col}{Character string: Name of the columns with prediction values.}

\item{init.stats}{Logical: Should the initial network stats be calculated? Default is FALSE.}

\item{final.stats}{Logical: Should the final network stats be calculated? Default is FALSE.}

\item{standard.set}{Dataframe with columns protein1, protein2 and another column with labels.}

\item{labels.col}{Character string: Name of the columns with labels.}

\item{rep.steps}{Integer: How many times should the algorithm be repeated before averaging the results? Default is 100.}

\item{SP.shift}{Numeric: By how much can the predictions be randomly shifted in SCBP algorithm?}

\item{SR.rows}{Numeric: Maximum by how many rows can a protein pair shift in order predictions in SR algorithm? Default is 20.}
}
\value{
A list with three elements: $data with columns 'protein1', 'protein2' and score contains all data for plotting or further processing
of the network, 'stats_initial' contains statistics of the network before clustering. and 'stats_final' contains statistics of the network after
clustering.
}
\description{
Using one of the network building algorithms, cluster initial network into individual protein complexes.
}
\examples{
X.refine.complexes(network.built)
}
