% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/X.scale.meltcurves.R
\name{X.scale.meltcurves}
\alias{X.scale.meltcurves}
\title{X.scale.meltcurves}
\usage{
X.scale.meltcurves(
  data = NULL,
  temperatures = c(37, 41, 45, 49, 53, 57, 61, 65, 69, 73)
)
}
\arguments{
\item{data}{Data frame with data containing a column 'id' and one 'Ab' column for each temperature.}

\item{temperatures}{Vector of integers: Temperatures used for the generation of the meltcurves.}
}
\value{
A list with two elements: $data with scaled data and $plot showing the data distribution before and after the scaling.
}
\description{
Scale measured protein abundances across temperature gradient.
}
\examples{
MCdata.scaled <- X.scale.meltcurves(MCdata.clean$data)
}
