% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Xmodel.tree.RF.R
\name{Xmodel.tree.RF}
\alias{Xmodel.tree.RF}
\title{Xmodel.tree.RF}
\usage{
Xmodel.tree.RF(
  data = NULL,
  mTry = 1,
  nodesize = 10,
  downsample = 1,
  rf.split = 5,
  ntree = 151
)
}
\arguments{
\item{data}{Data frame with pairwise differential values. Must contain columns "protein1" and "protein2", any number of columns with
predictors for modelling and a labels column with values 1 (for protein pairs that form a complex) and 0 (for protein pairs that
do not form a complex).}

\item{downsample}{Integer: How many times less of the non-interacting proteins should be used for the training? Default is 1. Applicable for C5.0 models.}

\item{eval.metric}{Character string: How should the model be evaluated in cross-validation? Defalt is "prc" for area under the precision-recall curve.
Other options are "roc" for area under the receiver-operator curve and "kappa" for Cohen's kappa.}

\item{CF}{Numeric vector: Complexity parameters to be cross-validated. Default is 0.1.}

\item{costs}{Integer: A number specifying how much is the cost of falsely predicting non-interacting protein pairs as interacting higher than vice versa.}
}
\value{
A list with two elements. $model contains the model and $predict.type contains a string that is used in
predict() to predict values using the model.
}
\description{
Random forest MAP-X models.
}
